+++ doc/changelog.rst
@@ -16,6 +16,7 @@ PyMongo 4.16 brings a number of changes including:
   Python 3.10+.  The minimum version is ``2.6.1`` to account for `CVE-2023-29483 <https://www.cve.org/CVERecord?id=CVE-2023-29483>`_.
 - Removed support for Eventlet.
   Eventlet is actively being sunset by its maintainers and has compatibility issues with PyMongo's dnspython dependency.
+- Added support for NumPy 1D-arrays in BSON Binary Vectors.
 
 Changes in Version 4.15.3 (2025/10/07)
 --------------------------------------